remoteproc: Introduce support for processor control
Modern System on Chip solutions are coming up with very
complex architectures and with multiple processing entities
on the same die, each entity being dedicated for its own specific
functionality.

In certain classes of products where safety is the major concern,
the controlling of other processing entities cannot be done directly
by the one running u-boot or HLOS. All the controlling capabilities
are abstracted and done by a centralized processor acting as master
of the SoC. This centralized processor has a complete picture on the
current status of the SoC. Typical example would be of TI's K3 generation
SoCs where SYSFW(M3) acts as a centralized master. For u-boot to load an
image to any other core, it should communicate with SYSFW.

In order to address these needs, a wrapper layer should exist
between remoreproc driver and the communication protocol to the
master core. This patch adds a wrapper layer using a new UCLASS_PROC_CONTROL
driver. All the services that are provided by the remoteproc driver
are available by this wrapper layer, so that any remoteproc driver
uses these services as necessary.

Sandbox tests to be added?

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>