am65x_evm: Add support to boot kernel from OSPI
Add support to boot Kernel out of OSPI. Setup the required environment
variables and sane defaults. Update the image offsets to match this
layout. Here is the flash layout(mtdparts env variable encapsulates this). Same
layout is passed to kernel as cmdline argument.

         0x0 +----------------------------+
             |     ospi.tiboot3(512K)     |
             |                            |
     0x80000 +----------------------------+
             |     ospi.tispl(2M)         |
             |                            |
    0x280000 +----------------------------+
             |     ospi.u-boot(5M)        |
             |                            |
    0x780000 +----------------------------+
             |     ospi.env(128K)         |
             |                            |
    0x7A0000 +----------------------------+
             |     Unused(384K)           |
             |                            |
    0x800000 +----------------------------+
             |     ospi.rootfs(UBIFS)     |
             |                            |
             +----------------------------+

Kernel Image and DT are expected to be present in the /boot folder of UBIFS
ospi.rootfs just like in SD card case. U-Boot looks for UBI volume named
"rootfs" for rootfs.

To boot from OSPI, at the U-Boot prompt:
=> setenv boot ubi
=> boot

Signed-off-by: Vignesh R <vigneshr@ti.com>