arm: K3: am654: Update board config for SYSFW 2018.06
Going from system firmware (SYSFW) version 2018.05a to 2018.06 a
change to the board configuration data structure that is provided to
SYSFW during startup was introduced that breaks compatibility with
the structure currently in U-Boot, causing the initial SPL startup
and with that the entire boot chain to fail silently.

Go ahead and update the board configuration structure according to
the new requirements. Note that with this patch applied, compatibility
with SYSFW 2018.05a (and earlier) versions will be broken.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>