board: ti: am654: Enable full 4GB of DDR available on EVM
In order to be able to use the full 4GB of DDR available on the AM654
EVM we need to make U-Boot aware of this so that the Kernel DTB file
can get updated accordingly. Since our boot process is split across
32-bit (R5) and 64-bit (A53) cores the simplest way to accomplish this
is to only configure the portion of U-Boot executing on the A53 for
the full amount of DDR memory available as follows:

Bank #0: 2 GiB @ 0x080000000
Bank #1: 2 GiB @ 0x880000000

Note that independent of the CPU core the DDR usage by U-Boot itself
is limited to first bank which comprises the lower memory DDR region.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>