mmc: add the MMC_CLK_ENABLE/DISABLE macro in mmc.h
commit 65117182543e58ba69f4abdc998423798137cbae upstream.

mmc_set_clock() function has the disable argument as bool type.
When mmc_set_clock is called, it might be passed to "true" or "false".
But it's too confusion whether clock is enabled or disabled with only
"true" and "false".
To prevent the confusion, replace to MMC_CLK_ENABLE/DISABLE macro from
true/false.

Signed-off-by: Jaehoon Chung <jh80.chung@samsung.com>
Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>