arm: K3: am654: Update board config for SYSFW 2018.07
Going from system firmware (SYSFW) version 2018.06 to 2018.07 several
changes to the board configuration process and the associated data
structures are introduced that break compatibility with the approach
and data structures currently used in U-Boot, causing the initial SPL
startup and with that the entire boot chain to fail silently.

To restore functionality go ahead and update the board configuration
process according to the new requirements by splitting it out into 4
distinct steps (general cfg, resource management, security, and power
management) while using the updated data structures. Do so in a single
commit as all this functionality is tightly coupled and must be used
as a single logical entity.

Note that the ABI now used with SYSFW 2018.07 is intended to be stable
and will be maintained in a backwards compatible fashion, so the plan
is to be able to introduce future SYSFW releases in a manner that won't
break clients such as U-Boot.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>