firmware: ti_sci: Make SYSFW version query part of probe
Starting with system firmware (SYSFW) version 2018.07 we are able to
query its version information early on. Take advantage of this by
moving the firmware version query via TISCI protocol into the driver's
probe function and remove the now redundant exposing of this function
from the public API and its invocations by consumers.

The added benefit is also that the TISCI probe function itself now
implicitly does a basic sanity and communication check with SYSFW
and with that has the chance to expose any possible issues earlier.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>