firmware: ti_sci: Rename misc to board ops to better reflect their use
Until now we have used a catch-all bucket called "misc ops" for TISCI
message related operations that did not warrant their own category. But
now with the TISCI firmware version operation removed from the API and
the proliferation of multiple board configuration related operations it
makes sense to go ahead and refactor the name of this set of functions
to "board ops" to better reflect what they are used for.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>