arm: K3: am654: Make SYSFW PM init the first of the config steps
The different system firmware (SYSFW) board configuration steps can be
performed in arbitrary order. Since the PM initialization directly
affects the device operating points (CPU frequency, etc.) and is also
a needed prerequisite for being able to start utilizing many peripheral
modules including the MAIN_UART0 which is currently used for the U-Boot
console it's a good idea to get this established and out of the way as
early on as possible. While at it, use this opportunity to move the
U-Boot console init and SYSW version message output a few steps up the
execution chain.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>