arm64: dts: k3-am6: Re-add sysreset controller node
An earlier commit f998e896e7e9 ("arm: dts: ti: Sync k3-am6* dts files
with Linux") removed the sysreset controller node from our DTS file,
which is a prerequisite for being able to trigger a device reset using
the U-Boot 'reset' command by way of SYSFW. Therefore go ahead and
restore this node to allow for this functionality to work.

Note that this assumes a SYSFW version is being used that actually
implements this functionality (not the case as of SYSFW 2018.07),
otherwise an error message is being output when attempting a reset.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>