configs: am65x_evm_r5: Enable early console in SPL
In order to allow diagnostic messages to be output pre and during System
Firmware (SYSFW) loading and startup enable the MCU_UART0 to be used as
an early UART during R5 SPL execution, until the MAIN_UART0 has been
brought up. This is also required to be able to perform UART-based boot
which is dependent on a functioning serial console of some sorts.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>