armV7R: K3: am654: Load SYSFW binary and config from boot media
In order to optimize and simplify the overall boot flow by avoiding the
double-authentication of the system firmware (SYSFW) binary image that
is currently being build directly into R5 SPL which also puts SPL right
at the size limit of what the ROM bootloader can handle as an initial
load go ahead and split SYSFW as well as the associated configuration
data from the R5 SPL image and instead load it from an image tree blob
named "sysfw.itb" from an FS-based boot media or from a specific offset
when using a memory device such as SPI flash.

To simplify the handling of and loading from the different boot media we
tap into the existing U-Boot SPL framework usually used for loading U-Boot
by building on an earlier commit that exposes some of that functionality.

Note that this initial implementation only supports the following boot
modes and renders all other boot modes non-operational:

- FS and RAW-based eMMC/SD card boot
- SPI Flash boot
- Y-Modem based serial boot

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>