arm: K3: Make U-Boot init hang on DDR failures
When the DDR initialization fails due to training failures and other
reasons an error message is being output to the U-Boot console. However
despite such failures obviously being very serious code execution is
allowed to continue which most likely will result in the device crashing
(and even if it doesn't immediately crash due to the nature of a specific
failure this will probably happen sometime later).

Improve this behavior by forcing U-Boot to hang so that the user is
alerted and forced to take action on any DDR init failures.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>