DM: GPIO: Fix da8xx GPIO indexing over GPIO 32
commit 1eddf549319541ae41b2c8f1cb7a01fe3b737b53 upstream

The GPIO banks are broken up into two 16-bit registers for each
bank set.  Unfortunately, the math that determines how to shift
blindly shifted by the number of the gpio.  This worked for gpio
numbers under 32, but higher gpio's are broken.  This fixes the
gpio index, so the bank is passed and the shift amount within
the register is passed now instead of the gpio number.

Fixes: 8e51c0f25406("dm: gpio: Add DM compatibility to
GPIO driver for Davinci")

Signed-off-by: Adam Ford <aford173@gmail.com>
Signed-off-by: Keerthy <j-keerthy@ti.com>