configs: am65x_evm_a53: Fix bootcommand for linux boot
check_android variable sets the mmc device to eMMC
and checks for android partitions. If partitions found, then
boots android else fallsback to linux boot. But default linux
boot is from SD card and the mmc device set at this point is
eMMC. So default linux boot is failing because of this setting.
Fix this by setting mmcdev to SD before calling envboot.

Fixes: 02e39656aa27f ("configs: am65x_evm: Run Android from eMMC")
Tested-by: Suman Anna <s-anna@ti.com>
Reviewed-by: Sam Protsenko <semen.protsenko@linaro.org>
Reported-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>