arch: arm: dts: DDR config settings for DDR4 @1600 MTS
This adds support for DDR4 running at 1600MTS (800MHZ) DDR Clock.

The dtsi file was generated using the TI DDR Config tool and used
directly without modification. The settings generated are sepecific
to the TI AM654x Base Board EVM. The tool output is included in
the k3-am654-r5-base-board.dts.

Please note the DDR clock settings to run the DDR at 800MHz are
defined in the memorycontroller node. The assinged-clock-rates
value is provided by the DDR config tool. The assigned-clocks
property is the Device ID and Clock ID of the k3_clks structure.

	assigned-clocks = <&k3_clks 20 1>;
	assigned-clock-rates = <DDR_PLL_FREQUENCY>;

Signed-off-by: Schuyler Patton <spatton@ti.com>
Signed-off-by: James Doublesin <doublesin@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>