am65x_evm: fastboot: Fix fastboot detection/trigger mechanism
The mechanism that checks if a fastboot should be performed by
testing the 'dofastboot' ENV variable for a value of 1 fails when
that ENV variable is not defined such as during a boot with the
initial/default environment, in which case the fastboot would
get triggered unintentionally.

Fix this by enclosing the dofastboot ENV variable in quotes so
that the test command evaluates as desired.

Fixes: 5a8d506f21d ("am65x_evm: fastboot: Enable reboot-bootloader")
Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>
Acked-by: Praneeth Bajjuri <praneeth@ti.com>
Reviewed-by: Sam Protsenko <semen.protsenko@linaro.org>