usb: udc-uclass: Fixed problem when no alias is defined in DT
commit 801f1fa44275 ("dm: usb: udc: Use SEQ_ALIAS to index the USB
gadget ports") changed the way the udevice if found. It uses the alias
to find a udevice for a given USB port number. In the commit log it was
stated that if no alias is provided, the bind order will be used
instead. However it doesn't work. Fixing this by adding a call to
uclass_get_device() if uclass_get_device_by_seq() fails.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Signed-off-by: Vignesh R <vigneshr@ti.com>