Revert "fs/fat: Re-use dynamic memory in case of SPL_SYS_MALLOC_SIMPLE"
This reverts commit 30f3e51731f31ae6f2bf281a3d2504c59afd97ec.

We no longer need this commit that was originally introduced to allow
R5 SPL on AM654 devices to load multiple files from FAT FS media in a
pre-SYSFW/pre-DDR environment with only very limited memory available.

Since this is also not something we want to push upstream reverting
this commit now will help reducing the amount of changes we carry.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>