Revert "ti_armv7_common: Skip default SPL malloc env setup on K3 R5"
This reverts commit 179f4490e4882aec9ff78015cc3ee891e74ce72b.

We no longer need this commit since we switched to using the full
malloc on R5 SPL on AM654 devices for which this commit was originally
introduced, an approach we intend to follow also moving forward. Hence
go ahead and remove the modifications.

Since this is also not something we want to push upstream reverting
this commit now will help reducing the amount of changes we carry.

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>