firmware: ti_sci: drop the device exclusive flag
When requesting a PM resource, the device exclusive flag locks the
resource to the currently running processor entity, preventing any
access from others. This leads into various problems with u-boot
framework; for example when SPL turns on A53 core for u-boot execution,
the A53 cluster + core are locked to R5 ownership, and there is no
way to unlock it, other than disabling the A53 core, and releasing the
resource. This obviously does not work out too well. Similar problem
is going to be with any remoteproc resource initialized with spl/u-boot,
as it is impossible to handover the resources to Linux side.

Signed-off-by: Tero Kristo <t-kristo@ti.com>