usb: ether: call _usb_eth_halt() if initialization fails
commit 1c1464c2b19f1b96885040b8adc9cf5ba3c7b32d upstream.

If the host does not respond in time, the initialization fails. However
the usb ether driver will still be registered. This will make
usb_gadget_probe_driver() fail the next time the initialization is
attempted because it cannot find an available UDC.

Fixing this by calling _usb_eth_halt() when the init fails.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>