configs: am335x_evm: enable DM_USB_GADGET and USB_ETHER in u-boot and SPL
commit 94f0d13c9bd7eca3f7f9369cf1f2f8f32f04097f upstream.

The AM335x ROM boot is able to download the SPL from a RNDIS connection
on USB0. To enable a full RNDIS boot flow (romboot -> SPL -> u-boot -> ..),
we can use USB_ETHER in SPL and u-boot.
However this increase the size of the SPL past its limit. So removing the
unused SPL_EXT_SUPPORT.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>