HACK: am65x: re-init mmc boot device after sysfw is loaded
Once sysfw is loaded, sysfw will configure all the PLLs to a known state.
During this time, clocks to the boot media gets reset and might observe a
glitch on the boot peripheral. Because of this boot peripheral might not
work if SPL continus to use the same settings. So, to avoid this, the boot
peripheral must be re initialized after loading sysfw.

Tested-by: Praneeth Bajjuri <praneeth@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>