ARM: mach-k3: Provide a FIT generator script
The FIT generator script is used to generated a ITS file that is, in turn,
used to produce a FIT image more complex than the default provided by
mkimage.
This generator produces a FIT image containing u-boot, the dtbs
in OF_LIST and all the DT overlays used by the board.
It gets the list of overlays by parsing the built-in.o object in the board
directory and extracting the strings ending in '.dtbo'.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>