armv7R: k3: sysfw-loader: Use spi memmapped addr when loading sysfw
In SPI bootmode, ROM initializes the SPI flash and access it in
memory mapped mode. For loading sysfw R5 SPL is currently loading the
image in the following manner:
- Initialize the SPI flash
- Copy sysfw to internal SRAM
- pass the copied buffer to DMSC ROM for starting sysfw on DMSC.

DMSC ROM then does copy the image to right location. Now the image
is copied twice and might be big hit for boot time. Since ROM enables
SPI flash in memory mapped mode, SPL can directly pass the memory
mapped pointer to ROM. With this ROM can directly copy the image
to the destination. Adding support for the same.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>