defconfig: k2x_hs: Fixes DHCP issue in u-boot
Fixes the inadvertent definition of
CONFIG_DTB_RESELECT and CONFIG_MULTI_DTB_FIT in
the K2x HS defconfigs, that happened as part of
u-boot migration from 2017 to 2018.

The inclusion of these config options causes
ethernet to stop working on K2x HS devices, as
they interfere with the installation of the secure
boot monitor

This patch also removes the above configs for the
K2 GP devices, as they are not needed, and to keep
the differences between the GP and HS defconfigs to
the minimum.

Signed-off-by: Madan Srinivas <madans@ti.com>